Conflicts:
	src/crypto/curve25519xsalsa20poly1305.rs
	src/crypto/xsalsa20poly1305.rs

+++ src/crypto/aes128ctr.rs
@@ -10,9 +10,10 @@ use ffi::{crypto_stream_aes128ctr,
 use libc::c_ulonglong;
 use std::intrinsics::volatile_set_memory;
 use std::iter::repeat;
+use std::ops::{Index, Range, RangeFrom, RangeFull, RangeTo};
 use randombytes::randombytes_into;
 
 stream_module!(crypto_stream_aes128ctr,
                crypto_stream_aes128ctr_xor,
-               crypto_stream_aes128ctr_KEYBYTES as usize,
-               crypto_stream_aes128ctr_NONCEBYTES as usize);
+               crypto_stream_aes128ctr_KEYBYTES,
+               crypto_stream_aes128ctr_NONCEBYTES);
