+++ src/crypto/hmacsha256.rs
@@ -9,14 +9,15 @@ use ffi::{crypto_auth_hmacsha256,
 };
 use libc::c_ulonglong;
 use std::intrinsics::volatile_set_memory;
+use std::ops::{Index, Range, RangeFrom, RangeFull, RangeTo};
 use randombytes::randombytes_into;
 use crypto::verify::verify_32;
 
 auth_module!(crypto_auth_hmacsha256,
              crypto_auth_hmacsha256_verify,
              verify_32,
-             crypto_auth_hmacsha256_KEYBYTES as usize,
-             crypto_auth_hmacsha256_BYTES as usize);
+             crypto_auth_hmacsha256_KEYBYTES,
+             crypto_auth_hmacsha256_BYTES);
 
 #[test]
 fn test_vector_1() {
