+++ src/crypto/salsa2012.rs
@@ -10,9 +10,10 @@ use ffi::{crypto_stream_salsa2012,
 use libc::c_ulonglong;
 use std::intrinsics::volatile_set_memory;
 use std::iter::repeat;
+use std::ops::{Index, Range, RangeFrom, RangeFull, RangeTo};
 use randombytes::randombytes_into;
 
 stream_module!(crypto_stream_salsa2012,
                crypto_stream_salsa2012_xor,
-               crypto_stream_salsa2012_KEYBYTES as usize,
-               crypto_stream_salsa2012_NONCEBYTES as usize);
+               crypto_stream_salsa2012_KEYBYTES,
+               crypto_stream_salsa2012_NONCEBYTES);
